Fix incorrect logic in getPSNUs
Where Countries double as PSNUs, they will not return in initial API call in getPSNUs. A separate API call is included to pull these countries separately and add them to the result of the first query.

Previous logic executed this conditional second query when there was more than 1 country doubling as a PSNU. This fix corrects to execute when there are more than 0 of these cases.