Update unPackSheet.R
If the first sheet with an error of a given type had only one error then the d[["tests"]][["#type#"]] object then this object is set up as a vector. If another sheet has multiple coulumn with errors of this type we get an error when we try add a vector for d[["tests"]][["#type#"]][[#sheet#]]. Initializing with an empty character vector will force each d[["tests"]][["#type#"]] object to be of type list even if the first error of a given type is attributable to only one column.