add checkDecimalValues.R
In keeping with the move of checking decimal values in the unPacksheet.R branch, I extended the checkDecimalValues.R function to account for unpacking SNUXIM if we end up needing to differentiate. I believe this validates the need for checkDecimalValues.R in the R directory since it would be used in at least two different functions in this package.