[DP-473]
In packageSetup as changed a list to a range just to make future updates 1 character change.

In packPSNUxIM Ofcourse tons of comments explaining the code. Made some small code updates to at least make it more readable.

I also compared this to packSNU and perhaps I don't understand how styleguide works, but it looks to me like neither are using that function as the graph on confluence states?