DP-916
Formula to calculate the Denominator (col T) needs to be reverted back to the old formula - =IF(SUM($P15)=0,"",ROUND(SUM($P15)*(1+$R15),0))

Conditional formatting rule that highlights yellow needs to only be applied to col T. Currently applied to col T and U. Col U only needs to flag red if U>T which is already a rule in place.

Add a note in Col F to end users. "Please refer to the User Guide on how to populate column F - K"